sched: ems: Correct cpu_util_wake() return value

+++ kernel/sched/ems/core.c
@@ -22,7 +22,7 @@ int task_util(struct task_struct *p)
 		return p->se.avg.util_avg;
 }
 
-int cpu_util_wake(int cpu, struct task_struct *p)
+unsigned long cpu_util_wake(int cpu, struct task_struct *p)
 {
 	struct cfs_rq *cfs_rq;
 	unsigned int util;
